完善强制退出steam客户端的逻辑

+++ CmdHandler.py
@@ -497,6 +497,16 @@ class CmdHandler(object):
 
     # 强制退出steam客户端
     def killSteamAppClient(self):
+        cmd = "ps -ef | grep -E '(steam.sh$|steam$)' | grep -v grep | awk '{print $2}' | wc -l"
+        result = os.popen(cmd).read()
+        # steam客户端进程未启动
+        if result.strip() == "0":
+            return {
+                "cmdCode": 0,
+                "result": None,
+                "errMsg": "",
+            }
+
         command = "ps -ef | grep -E '(steam.sh$|steam$)' | grep -v grep | awk '{print $2}' | xargs kill -9"
         return runShellCommand(command)
